pyproject.toml
[project]
name = "telegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "markdown2>=2.5.4",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
